backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
requests==2.32.3
gunicorn==22.0.0
Pillow==10.4.0
requests==2.32.3